+++ extensions/jsr/src/components/ListItem.tsx
@@ -12,15 +12,17 @@ type ListItemProps = {
   item: SearchResultDocument;
   toggleDetails: () => void;
   isShowingDetails: boolean;
+  extraActions: JSX.Element;
 };
 
-const ListItem = ({ item, toggleDetails, isShowingDetails }: ListItemProps) => {
+const ListItem = ({ item, toggleDetails, isShowingDetails, extraActions }: ListItemProps) => {
   const progress = item.score ?? 0;
   const iconColor = progress >= 80 ? Color.Green : progress >= 50 ? Color.Yellow : Color.Red;
   const icons = compatIcons(item);
 
   return (
     <List.Item
+      id={`${item.scope}/${item.name}`}
       title={item.id}
       subtitle={item.description}
       accessories={
@@ -34,26 +36,34 @@ const ListItem = ({ item, toggleDetails, isShowingDetails }: ListItemProps) => {
           {preferences.openWebsiteByDefault ? (
             <>
               <Action.OpenInBrowser
-                title="Open in JSR"
+                title="Open Main Page (JSR)"
                 icon={{ source: "jsr.svg" }}
                 url={`https://jsr.io/${item.id}`}
               />
-              <Action
-                title="Toggle Details"
-                icon={Icon.AppWindowSidebarLeft}
-                onAction={() => toggleDetails()}
-                shortcut={{ key: "o", modifiers: ["cmd"] }}
+              <Action title="Toggle Details" icon={Icon.AppWindowSidebarLeft} onAction={() => toggleDetails()} />
+              <Action.OpenInBrowser
+                title="Open Docs (JSR)"
+                icon={{ source: "jsr.svg" }}
+                url={`https://jsr.io/${item.id}/doc`}
+                shortcut={{ key: "enter", modifiers: ["cmd", "shift"] }}
               />
+              {extraActions}
             </>
           ) : (
             <>
               <Action title="Toggle Details" onAction={() => toggleDetails()} icon={Icon.AppWindowSidebarLeft} />
               <Action.OpenInBrowser
-                title="Open in JSR"
+                title="Open Main Page (JSR)"
                 icon={{ source: "jsr.svg" }}
                 url={`https://jsr.io/${item.id}`}
-                shortcut={{ key: "o", modifiers: ["cmd"] }}
               />
+              <Action.OpenInBrowser
+                title="Open Docs (JSR)"
+                icon={{ source: "jsr.svg" }}
+                url={`https://jsr.io/${item.id}/doc`}
+                shortcut={{ key: "enter", modifiers: ["cmd", "shift"] }}
+              />
+              {extraActions}
             </>
           )}